feat: different audio visualizer, redesign react demo
