+++ src/demo/src/App.module.css
@@ -1,23 +1,26 @@
+.header,
+.footer {
+  pointer-events: none;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  position: sticky;
+  background: rgba(0, 0, 0, 0);
+  width: 100%;
+}
+
 .header {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    position: sticky;
-    background: black;
-    top: 0;
-    width: 100%;
-    margin-bottom: 5px;
-    box-shadow: 0 8px 10px 5px rgba(0, 0, 0, 20);
+  top: 0;
+  margin-bottom: calc(1px + 2vmin);
 }
 
 .footer {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    position: sticky;
-    background: black;
-    bottom: 0;
-    width: 100%;
-    margin-top: 5px;
-    box-shadow: 0 -5px 10px 5px rgba(0, 0, 0, 20);
+  bottom: 0;
+}
+
+.spectrum {
+  position: absolute;
+  position: fixed;
+  z-index: -5;
+  height: 100%;
 }
